社内掲示板アプリ　billboardsテーブルにファイル情報を格納
