+++ src/main/java/jp/sys_link/action/BillboardsAction.java
@@ -11,12 +11,9 @@ import javax.servlet.http.HttpServletRequest;
 
 import jp.sys_link.entity.Billboards;
 import jp.sys_link.entity.NameMst;
-import jp.sys_link.entity.Upfile;
 import jp.sys_link.form.BillboardsForm;
-import jp.sys_link.form.UpfileForm;
 import jp.sys_link.service.BillboardsService;
 import jp.sys_link.service.NameMstService;
-import jp.sys_link.service.UpfileService;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.struts.upload.FormFile;
@@ -31,21 +28,13 @@ public class BillboardsAction {
 	public List<Billboards> billboardsItems;
 	public List<NameMst> nameMstItems;
 	public Billboards billboardItems;
-	public List<Upfile> upfileItems;
 
 	@ActionForm
 	@Resource
 	private BillboardsForm billboardsForm;
 
-	@ActionForm
-	@Resource
-	private UpfileForm upfileForm;
-
 	@Resource
 	private BillboardsService billboardsService;
-
-	@Resource
-	private UpfileService upfileService;
 
 	@Resource
 	private NameMstService nameMstService;
@@ -93,13 +82,13 @@ public class BillboardsAction {
 		// TODO ログイン機能を追加することになったら、セッションからログインIDを取得する
 		billboardsForm.setUserId("1");
 
+		// TODO アップロード処理
+		upload(billboardsForm.getFormFile());
+
 		Billboards entity = Beans
 				.createAndCopy(Billboards.class, billboardsForm)
 				.dateConverter("yyyy/MM/dd").execute();
 		billboardsService.insert(entity);
-
-		// TODO アップロード処理
-		upload(billboardsForm.getFormFile());
 
 		return "/billboards/";
 	}
@@ -120,35 +109,19 @@ public class BillboardsAction {
 				.dateConverter("yyyy-MM-dd").execute();
 		billboardsService.delete(entity);
 
-		// TODO 上記で削除したIDが持っていたファイル情報の削除
-		upfileItems = jdbcManager.from(Upfile.class)
-				.where("billboardId = ?", entity.getId()).getResultList();
-		for (Upfile upfileEntity : upfileItems)
-			upfileService.delete(upfileEntity);
-
 		return "/billboards/";
 	}
 
 	protected void upload(FormFile file) {
-		/**
-		 * アップロードできているかの確認。 String path = application.getRealPath("/WEB-INF/" +
-		 * file.getFileName()); UploadUtil.write(path, file);
-		 */
 
 		// TODO アップロードしたファイル情報をデータベースに格納
 		// アップロードされたかどうか、ファイルネームで判定
 		if (file.getFileName() != "") {
 
-			Billboards entity = billboardsService.findByMaxId();
 			try {
 				InputStream is = file.getInputStream();
-				upfileForm.setFile(IOUtils.toByteArray(is));
-				upfileForm.setBillboardId(entity.getId().toString());
-				upfileForm.setFileName(file.getFileName());
-
-				Upfile upfile = Beans.createAndCopy(Upfile.class, upfileForm)
-						.execute();
-				upfileService.insert(upfile);
+				billboardsForm.setFile(IOUtils.toByteArray(is));
+				billboardsForm.setFileName(file.getFileName());
 
 			} catch (FileNotFoundException e) {
 				e.printStackTrace();
