+++ src/main/java/jp/sys_link/form/BillboardsForm.java
@@ -24,6 +24,10 @@ public class BillboardsForm {
 
 	private FormFile[] formFiles;
 
+	private String fileName = "";
+
+	private byte[] file;
+
 	public String getId() {
 		return id;
 	}
@@ -86,6 +90,22 @@ public class BillboardsForm {
 
 	public void setFormFiles(FormFile[] formFiles) {
 		this.formFiles = formFiles;
+	}
+
+	public String getFileName() {
+		return fileName;
+	}
+
+	public void setFileName(String fileName) {
+		this.fileName = fileName;
+	}
+
+	public byte[] getFile() {
+		return file;
+	}
+
+	public void setFile(byte[] file) {
+		this.file = file;
 	}
 
 }
