--- src/main/java/jp/sys_link/service/UpfileService.java
@@ -1,14 +1,0 @@
-package jp.sys_link.service;
-
-import javax.annotation.Generated;
-
-import jp.sys_link.entity.Upfile;
-
-/**
- * {@link Upfile}のサービスクラスです。
- *
- */
-@Generated(value = { "S2JDBC-Gen 2.4.45","org.seasar.extension.jdbc.gen.internal.model.ServiceModelFactoryImpl" }, date = "2015/05/26 14:06:39")
-public class UpfileService extends AbstractService<Upfile> {
-
-}
