--- src/test/java/jp/sys_link/service/UpfileServiceTest.java
@@ -1,34 +1,0 @@
-package jp.sys_link.service;
-
-import javax.annotation.Generated;
-import org.seasar.extension.unit.S2TestCase;
-
-/**
- * {@link UpfileService}のテストクラスです。
- * 
- */
-@Generated(value = {"S2JDBC-Gen 2.4.45", "org.seasar.extension.jdbc.gen.internal.model.ServiceTestModelFactoryImpl"}, date = "2015/05/26 14:06:42")
-public class UpfileServiceTest extends S2TestCase {
-
-    private UpfileService upfileService;
-
-    /**
-     * 事前処理をします。
-     * 
-     * @throws Exception
-     */
-    @Override
-    protected void setUp() throws Exception {
-        super.setUp();
-        include("app.dicon");
-    }
-
-    /**
-     * {@link #upfileService}が利用可能であることをテストします。
-     * 
-     * @throws Exception
-     */
-    public void testAvailable() throws Exception {
-        assertNotNull(upfileService);
-    }
-}
